Appli/ml-core/requirements.txt
opencv-python
ultralytics
face_recognition
supabase
numpy
requests
python-dotenv
psycopg2-binary
flask
flask-cors
